+++ api/api/src/main/kotlin/nl/tno/federated/api/event/mapper/EventType.kt
@@ -1,6 +1,0 @@
-package nl.tno.federated.api.event.mapper
-
-
-import jakarta.validation.constraints.NotNull
-
-data class EventType(@NotNull val eventType: String, @NotNull val rml: String, val shacl: String?)