+++ api/api/src/main/kotlin/nl/tno/federated/api/event/type/EventType.kt
@@ -1,0 +1,6 @@
+package nl.tno.federated.api.event.type
+
+
+import jakarta.validation.constraints.NotNull
+
+data class EventType(@NotNull val eventType: String, @NotNull val rml: String, val shacl: String?, val schemaDefinition: String? = null)