+++ api/api/src/main/kotlin/nl/tno/federated/api/event/type/EventTypeMappingException.kt
@@ -1,0 +1,3 @@
+package nl.tno.federated.api.event.type
+
+class EventTypeMappingException(message: String) : Exception(message)