+++ api/api/src/main/kotlin/nl/tno/federated/api/event/type/EventTypeRepository.kt
@@ -1,4 +1,4 @@
-package nl.tno.federated.api.event
+package nl.tno.federated.api.event.type
 
 import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository